redesign to full page dropzone
extract devtools to its own component but comment out for now
