+++ apps/architect-vite/src/components/ExternalLink.tsx
@@ -16,11 +16,7 @@ const ExternalLink = ({ children, href }: ExternalLinkProps) => {
 	};
 
 	return (
-		<a
-			href={href}
-			onClick={handleClick}
-			className="font-bold underline decoration-[color:var(--color-action)] underline-offset-4 decoration-2"
-		>
+		<a href={href} onClick={handleClick} className="action-link">
 			{children}
 		</a>
 	);
